feat: move tests to test directory
