+++ test/features/sendSummaryComment.test.js
