+++ test/lcov.test.js
