+++ test/services/barecheckApi.test.js
@@ -1,0 +1,165 @@
+const sinon = require("sinon");
+const { assert } = require("chai");
+
+const { importMock } = require("../utils");
+
+const path = "services/barecheckApi";
+
+const defaultStubValues = {
+  axios: {}
+};
+
+const barecheckApiMock = importMock(path, defaultStubValues, ({ axios }) => ({
+  axios
+}));
+
+describe(path, () => {
+  describe("createGithubAccessToken", () => {
+    it("should call axios post with proper params", async () => {
+      const githubAppToken = "github-token:123";
+      const mutationResponse = {
+        data: {
+          data: {
+            createGithubAccessToken: {
+              success: true
+            }
+          }
+        }
+      };
+      const axios = {
+        post: sinon.stub().returns(mutationResponse)
+      };
+
+      const { createGithubAccessToken } = barecheckApiMock({
+        axios
+      });
+
+      await createGithubAccessToken(githubAppToken);
+
+      assert.isTrue(axios.post.calledOnce);
+      const [requestPath, body, headers] = axios.post.firstCall.args;
+      assert.deepEqual(requestPath, "https://api.barecheck.com/graphql");
+      assert.deepEqual(body.variables, { githubAppToken });
+      assert.deepEqual(headers, {
+        headers: {
+          "Content-Type": "application/json"
+        }
+      });
+    });
+
+    it("should return data with token from response", async () => {
+      const githubAppToken = "github-token:123";
+      const token = "1234";
+
+      const mutationResponse = {
+        data: {
+          data: {
+            createGithubAccessToken: {
+              success: true,
+              token
+            }
+          }
+        }
+      };
+      const axios = {
+        post: sinon.stub().returns(mutationResponse)
+      };
+
+      const { createGithubAccessToken } = barecheckApiMock({
+        axios
+      });
+
+      const res = await createGithubAccessToken(githubAppToken);
+
+      assert.deepEqual(res, mutationResponse.data.data.createGithubAccessToken);
+    });
+
+    it("should throw error once mutation doesn't have succes:True", async () => {
+      const githubAppToken = "github-token:123";
+
+      const mutationResponse = {
+        data: {
+          data: {
+            createGithubAccessToken: {
+              success: false
+            }
+          }
+        }
+      };
+      const axios = {
+        post: sinon.stub().returns(mutationResponse)
+      };
+
+      const { createGithubAccessToken } = barecheckApiMock({
+        axios
+      });
+
+      try {
+        await createGithubAccessToken(githubAppToken);
+        assert.fail("createGithubAccessToken should throw an error");
+      } catch {
+        assert.isTrue(axios.post.calledOnce);
+      }
+    });
+
+    it("should throw error once there is no data in the response", async () => {
+      const githubAppToken = "github-token:123";
+
+      const mutationResponse = {
+        errors: ["some error"]
+      };
+      const axios = {
+        post: sinon.stub().returns(mutationResponse)
+      };
+
+      const { createGithubAccessToken } = barecheckApiMock({
+        axios
+      });
+
+      try {
+        await createGithubAccessToken(githubAppToken);
+        assert.fail("createGithubAccessToken should throw an error");
+      } catch {
+        assert.isTrue(axios.post.calledOnce);
+      }
+    });
+  });
+
+  describe("setProjectMetric", () => {
+    it("should call axios post with proper params", async () => {
+      const apiKey = "api-key";
+      const branch = "master";
+      const commit = "e2e2e2";
+      const coverage = 93;
+
+      const mutationResponse = {
+        data: {
+          data: {
+            setProjectMetric: {
+              success: true
+            }
+          }
+        }
+      };
+      const axios = {
+        post: sinon.stub().returns(mutationResponse)
+      };
+
+      const { setProjectMetric } = barecheckApiMock({
+        axios
+      });
+
+      await setProjectMetric(apiKey, branch, commit, coverage);
+
+      assert.isTrue(axios.post.calledOnce);
+      const [requestPath, body, headers] = axios.post.firstCall.args;
+      assert.deepEqual(requestPath, "https://api.barecheck.com/graphql");
+      assert.deepEqual(body.variables, { apiKey, branch, commit, coverage });
+      assert.deepEqual(headers, {
+        headers: {
+          "Content-Type": "application/json"
+        }
+      });
+    });
+  });
+});
